+++ examples/ipfs-kad/src/main.rs
@@ -20,16 +20,16 @@
 
 #![doc = include_str!("../README.md")]
 
+use std::iter;
 use std::num::NonZeroUsize;
 use std::ops::Add;
 use std::time::{Duration, Instant};
-use std::iter;
 
 use anyhow::{bail, Result};
 use clap::Parser;
 use futures::StreamExt;
+use libp2p::swarm::{StreamProtocol, SwarmEvent};
 use libp2p::{bytes::BufMut, identity, kad, noise, tcp, yamux, PeerId};
-use libp2p::swarm::{SwarmEvent, StreamProtocol};
 use tracing_subscriber::EnvFilter;
 
 const BOOTNODES: [&str; 4] = [